Подсказка "Организации"

+++ 10Model/Customer/Organization.cs
@@ -8,14 +8,14 @@ namespace _10Model.Customer
         public string NameFullOpf { get; set; } //Полное наименование
         public string NameShortOpf { get; set; } //Сокращенное наименование
         public string Opf { get; set; } //Организационно-правовая форма
-        public long Ogrn { get; set; } //ОГРН
+        public ulong Ogrn { get; set; } //ОГРН
         public DateTime OgrnDate { get; set; } //Дата регистрации
-        public int Inn { get; set; } //ИНН
-        public int Kpp { get; set; } //КПП
+        public ulong Inn { get; set; } //ИНН
+        public ulong Kpp { get; set; } //КПП
         public string Bank { get; set; } //Название банка
-        public int Bik { get; set; } //БИК Банка
-        public long PayAccount { get; set; } //Расчетный счет
-        public long CorrAccount { get; set; } //Корреспондентский счет
+        public ulong Bik { get; set; } //БИК Банка
+        public ulong PayAccount { get; set; } //Расчетный счет
+        public ulong CorrAccount { get; set; } //Корреспондентский счет
         public int DirectorId { get; set; }
         public virtual Director Director { get; set; }
         public virtual Address AddressRegistration { get; set; }
